refactor: RegisterController 분리
